Initial working version
I re-implemented the decoder by Christian, stripping away all the
things that are not needed for the other encoding schemes (that we
don't need).  I also put all of the conversions indices -> doubles in
the class MuonCandidate. I believe that this part will not be needed
within Simon's monitoring framework, and this seems like a reasonable
way to factor this out.
Now I need to add unit tests.