fix array index
This wasn't clear from the initial description.
Now using the side bit in conjunction to the unsigned eta value to map
the signed eta index to the first array index as clarified by Marek.

Note to self: use `check_lut_dimension.py` to check the dimention of
the LUT, and the rangoe of the first index:
eta:  (16, 7, 8)
phi:  (16, 7, 8)
pt:  (6,)

There is still one thing that I don't understand (ask Marek): the
first index (~octant) is packed in 1b (side) + 3b (octant). So with 4b
it should be in the range 0-15.
However, in the file
svn.cern.ch/reps/atl1calo/firmware/L1Topo/trunk/processor/sources/common/realTimePath/realTimeDataPath.vhd
there is an extra /2 in
  MuonEtaTab(i/2+j*8)
which I don't need in my c++...to be followed up.